coefficients.jl

Astrometry.SOFA.fad03 — Function

fad03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean elongation of the Moon from the Sun.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

D – D, radians (Note 2)

Note

Though Δt is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and is from Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Astrometry.SOFA.fae03 — Function

fae03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean longitude of Earth.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

mlon – mean longitude of Earth, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and comes from Souchay et al. (1999) after Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Souchay, J., Loysel, B., Kinoshita, H., Folgueira, M. 1999, Astron.Astrophys.Supp.Ser. 135, 111

Astrometry.SOFA.faf03 — Function

faf03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean longitude of the Moon minus mean longitude of the ascending node.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

F – F, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and is from Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Astrometry.SOFA.faju03 — Function

faju03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean longitude of Jupiter.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

mlon – mean longitude of Jupiter, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and comes from Souchay et al. (1999) after Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Souchay, J., Loysel, B., Kinoshita, H., Folgueira, M. 1999, Astron.Astrophys.Supp.Ser. 135, 111

Astrometry.SOFA.fal03 — Function

fal03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean anomaly of the Moon.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

l – l, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and is from Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Astrometry.SOFA.falp03 — Function

falp03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean anomaly of the Sun.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

l – l', radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and is from Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Astrometry.SOFA.fama03 — Function

fama03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean longitude of Mars.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

mlon – mean longitude of Mars, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and comes from Souchay et al. (1999) after Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Souchay, J., Loysel, B., Kinoshita, H., Folgueira, M. 1999, Astron.Astrophys.Supp.Ser. 135, 111

Astrometry.SOFA.fame03 — Function

fame03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean longitude of Mercury.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

mlon – mean longitude of Mercury, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and comes from Souchay et al. (1999) after Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Souchay, J., Loysel, B., Kinoshita, H., Folgueira, M. 1999, Astron.Astrophys.Supp.Ser. 135, 111

Astrometry.SOFA.fane03 — Function

fane03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean longitude of Neptune.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

mlon – mean longitude of Neptune, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and is adapted from Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Astrometry.SOFA.faom03 — Function

faom03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean longitude of the Moon's ascending node.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

Ω – Ω, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and is from Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J., 1994, Astron.Astrophys. 282, 663-683.

Astrometry.SOFA.fapa03 — Function

fapa03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): general accumulated precession in longitude.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

long – general precession in longitude, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003). It is taken from Kinoshita & Souchay (1990) and comes originally from Lieske et al. (1977).

References

Kinoshita, H. and Souchay J. 1990, Celest.Mech. and Dyn.Astron. 48, 187

Lieske, J.H., Lederle, T., Fricke, W. & Morando, B. 1977, Astron.Astrophys. 58, 1-16

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Astrometry.SOFA.fasa03 — Function

fasa03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean longitude of Saturn.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

mlon – mean longitude of Saturn, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and comes from Souchay et al. (1999) after Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Souchay, J., Loysel, B., Kinoshita, H., Folgueira, M. 1999, Astron.Astrophys.Supp.Ser. 135, 111

Astrometry.SOFA.faur03 — Function

faur03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean longitude of Uranus.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

mlon – mean longitude of Uranus, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and is adapted from Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Astrometry.SOFA.fave03 — Function

fave03(Δt::AbstractFloat)

Fundamental argument, IERS Conventions (2003): mean longitude of Venus.

Input

Δt – TDB, Julian centuries since J2000.0 (Note 1)

Output

mlon – mean longitude of Venus, radians (Note 2)

Note

Though t is strictly TDB, it is usually more convenient to use TT, which makes no significant difference.

The expression used is as adopted in IERS Conventions (2003) and comes from Souchay et al. (1999) after Simon et al. (1994).

References

McCarthy, D. D., Petit, G. (eds.), IERS Conventions (2003), IERS Technical Note No. 32, BKG (2004)

Simon, J.-L., Bretagnon, P., Chapront, J., Chapront-Touze, M., Francou, G., Laskar, J. 1994, Astron.Astrophys. 282, 663-683

Souchay, J., Loysel, B., Kinoshita, H., Folgueira, M. 1999, Astron.Astrophys.Supp.Ser. 135, 111